+++ main.tf
@@ -1,15 +1,56 @@
 locals {
   aws_alb_ingress_controller_docker_image = "docker.io/amazon/aws-alb-ingress-controller:v${var.aws_alb_ingress_controller_version}"
+  aws_alb_ingress_controller_version      = var.aws_alb_ingress_controller_version
   aws_alb_ingress_class                   = "alb"
+  aws_vpc_id                              = data.aws_vpc.selected.id
+  aws_region_name                         = data.aws_region.current.name
+}
+
+data "aws_vpc" "selected" {
+  id = var.k8s_cluster_type == "eks" ? data.aws_eks_cluster.selected[0].vpc_config[0].vpc_id : var.aws_vpc_id
+}
+
+data "aws_region" "current" {
+  name = var.aws_region_name
+}
+
+data "aws_caller_identity" "current" {}
+
+# The EKS cluster (if any) that represents the installation target.
+data "aws_eks_cluster" "selected" {
+  count = var.k8s_cluster_type == "eks" ? 1 : 0
+  name  = var.k8s_cluster_name
 }
 
 data "aws_iam_policy_document" "ec2_assume_role" {
+  count = var.k8s_cluster_type == "vanilla" ? 1 : 0
   statement {
     actions = ["sts:AssumeRole"]
 
     principals {
       type        = "Service"
       identifiers = ["ec2.amazonaws.com"]
+    }
+  }
+}
+
+data "aws_iam_policy_document" "eks_oidc_assume_role" {
+  count = var.k8s_cluster_type == "eks" ? 1 : 0
+  statement {
+    actions = ["sts:AssumeRoleWithWebIdentity"]
+    effect  = "Allow"
+    condition {
+      test     = "StringEquals"
+      variable = "${replace(data.aws_eks_cluster.selected[0].identity[0].oidc[0].issuer, "https://", "")}:sub"
+      values = [
+        "system:serviceaccount:${var.k8s_namespace}:aws-alb-ingress-controller"
+      ]
+    }
+    principals {
+      identifiers = [
+        "arn:aws:iam::${data.aws_caller_identity.current.account_id}:oidc-provider/${replace(data.aws_eks_cluster.selected[0].identity[0].oidc[0].issuer, "https://", "")}"
+      ]
+      type = "Federated"
     }
   }
 }
@@ -23,7 +64,7 @@ resource "aws_iam_role" "this" {
 
   force_detach_policies = true
 
-  assume_role_policy = data.aws_iam_policy_document.ec2_assume_role.json
+  assume_role_policy = var.k8s_cluster_type == "vanilla" ? data.aws_iam_policy_document.ec2_assume_role[0].json : data.aws_iam_policy_document.eks_oidc_assume_role[0].json
 }
 
 data "aws_iam_policy_document" "alb_management" {
@@ -144,7 +185,7 @@ data "aws_iam_policy_document" "alb_management" {
 
 resource "aws_iam_policy" "this" {
   name        = "${var.aws_resource_name_prefix}${var.k8s_cluster_name}-alb-management"
-  description = "Permissions that are required to manage the AWS Application Load Balancer."
+  description = "Permissions that are required to manage AWS Application Load Balancers."
   path        = var.aws_iam_path_prefix
   policy      = data.aws_iam_policy_document.alb_management.json
 }
@@ -159,7 +200,11 @@ resource "kubernetes_service_account" "this" {
   metadata {
     name      = "aws-alb-ingress-controller"
     namespace = var.k8s_namespace
-
+    annotations = {
+      # This annotation is only used when running on EKS which can
+      # use IAM roles for service accounts.
+      "eks.amazonaws.com/role-arn" = aws_iam_role.this.arn
+    }
     labels = {
       "app.kubernetes.io/name"       = "aws-alb-ingress-controller"
       "app.kubernetes.io/managed-by" = "terraform"
@@ -257,7 +302,7 @@ resource "kubernetes_deployment" "this" {
 
     labels = {
       "app.kubernetes.io/name"       = "aws-alb-ingress-controller"
-      "app.kubernetes.io/version"    = local.aws_alb_ingress_controller_version
+      "app.kubernetes.io/version"    = "v${local.aws_alb_ingress_controller_version}"
       "app.kubernetes.io/managed-by" = "terraform"
     }
 
@@ -281,9 +326,10 @@ resource "kubernetes_deployment" "this" {
           "app.kubernetes.io/name"    = "aws-alb-ingress-controller"
           "app.kubernetes.io/version" = local.aws_alb_ingress_controller_version
         }
-
         annotations = {
-          # Annotation to be used by KIAM
+          # Annotation which is only used by KIAM and kube2iam.
+          # Should be ignored by your cluster if using IAM roles for service accounts, e.g.
+          # when running on EKS.
           "iam.amazonaws.com/role" = aws_iam_role.this.arn
         }
       }
@@ -301,8 +347,8 @@ resource "kubernetes_deployment" "this" {
           args = [
             "--ingress-class=${local.aws_alb_ingress_class}",
             "--cluster-name=${var.k8s_cluster_name}",
-            "--aws-vpc-id=${var.aws_vpc_id}",
-            "--aws-region=${var.aws_region_name}",
+            "--aws-vpc-id=${local.aws_vpc_id}",
+            "--aws-region=${local.aws_region_name}",
             "--aws-max-retries=10",
           ]